kernel-modules-headers: Add libyaml dependency
The kernel build introduced a dependeny on libyaml on v4.20 that checks whether
/usr/include/yaml.h exists. If it does, it tries to use it on the dtc
build.

https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=f858927fd6ce394a7f431153d44ad0a09e8f49a1

The check for /usr/include/yaml.h is a host contamination, so the effect
is that the build works if the host machine does not have libyaml
installed and fails if it does with:

scripts/dtc/yamltree.c:23:10: fatal error: yaml.h: No such file or
directory

Adding the explicit dependency on libyaml will make it work for both
cases even if it's unneeded in the case where the host machine does not
contain /usr/include/yaml.h.

Patching the kernel source to avoid this cross-contamination is a
different alternative, but the choice between maintaining a kernel patch
or a dependency addition on this package which is already owned my
meta-balena has made me decide for the latter.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>